deps/v8/test/webkit/fast/js/Promise-reject-expected.txt

Test Promise rejection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS promiseState is "pending"
PASS promiseState is "pending"
PASS promiseState is "rejected"
PASS promiseResult is "hello"
PASS successfullyParsed is true

TEST COMPLETE
